+++ aiodocker/__init__.py
@@ -2,7 +2,7 @@ from .docker import Docker
 from .exceptions import DockerContainerError, DockerError
 
 
-__version__ = "0.19.1"
+__version__ = "0.21.0"
 
 
 __all__ = ("Docker", "DockerError", "DockerContainerError")