+++ aiodocker/secrets.py
@@ -1,15 +1,15 @@
 import json
 from base64 import b64encode
-from typing import Any, List, Mapping
+from typing import Any, List, Mapping, Optional
 
 from .utils import clean_filters, clean_map
 
 
-class DockerSecrets(object):
+class DockerSecrets:
     def __init__(self, docker):
         self.docker = docker
 
-    async def list(self, *, filters: Mapping = None) -> List[Mapping]:
+    async def list(self, *, filters: Optional[Mapping] = None) -> List[Mapping]:
         """
         Return a list of secrets
 
@@ -33,9 +33,9 @@ class DockerSecrets(object):
         data: str,
         *,
         b64: bool = False,
-        labels: List = None,
-        driver: Mapping = None,
-        templating: Mapping = None,
+        labels: Optional[Mapping[str, str]] = None,
+        driver: Optional[Mapping] = None,
+        templating: Optional[Mapping] = None,
     ) -> Mapping[str, Any]:
         """
         Create a secret
@@ -82,9 +82,7 @@ class DockerSecrets(object):
             a dict with info about a secret
         """
 
-        response = await self.docker._query_json(
-            "secrets/{secret_id}".format(secret_id=secret_id), method="GET"
-        )
+        response = await self.docker._query_json(f"secrets/{secret_id}", method="GET")
         return response
 
     async def delete(self, secret_id: str) -> bool:
@@ -98,9 +96,7 @@ class DockerSecrets(object):
             True if successful
         """
 
-        async with self.docker._query(
-            "secrets/{secret_id}".format(secret_id=secret_id), method="DELETE"
-        ):
+        async with self.docker._query(f"secrets/{secret_id}", method="DELETE"):
             return True
 
     async def update(
@@ -108,12 +104,12 @@ class DockerSecrets(object):
         secret_id: str,
         version: str,
         *,
-        name: str = None,
-        data: str = None,
+        name: Optional[str] = None,
+        data: Optional[str] = None,
         b64: bool = False,
-        labels: List = None,
-        driver: Mapping = None,
-        templating: Mapping = None,
+        labels: Optional[Mapping[str, str]] = None,
+        driver: Optional[Mapping] = None,
+        templating: Optional[Mapping] = None,
     ) -> bool:
         """
         Update a secret.
@@ -155,7 +151,7 @@ class DockerSecrets(object):
         request_data = json.dumps(clean_map(spec))
 
         await self.docker._query_json(
-            "secrets/{secret_id}/update".format(secret_id=secret_id),
+            f"secrets/{secret_id}/update",
             method="POST",
             data=request_data,
             params=params,