+++ aiodocker/swarm.py
@@ -1,22 +1,22 @@
-from typing import Iterable, Mapping
+from typing import Iterable, Mapping, Optional
 
 from .utils import clean_map
 
 
-class DockerSwarm(object):
+class DockerSwarm:
     def __init__(self, docker) -> None:
         self.docker = docker
 
     async def init(
         self,
         *,
-        advertise_addr: str = None,
-        data_path_port: int = None,
-        default_addr_pool: list = None,
+        advertise_addr: Optional[str] = None,
+        data_path_port: Optional[int] = None,
+        default_addr_pool: Optional[list] = None,
         listen_addr: str = "0.0.0.0:2377",
         force_new_cluster: bool = False,
         subnet_size: int = 24,
-        swarm_spec: Mapping = None,
+        swarm_spec: Optional[Mapping] = None,
     ) -> str:
         """
         Initialize a new swarm.
@@ -66,8 +66,8 @@ class DockerSwarm(object):
         remote_addrs: Iterable[str],
         listen_addr: str = "0.0.0.0:2377",
         join_token: str,
-        advertise_addr: str = None,
-        data_path_addr: str = None,
+        advertise_addr: Optional[str] = None,
+        data_path_addr: Optional[str] = None,
     ) -> bool:
         """
         Join a swarm.