+++ aiodocker/utils.py
@@ -1,15 +1,12 @@
-import asyncio
 import base64
 import codecs
 import json
-import sys
 import tarfile
 import tempfile
 from io import BytesIO
 from typing import (
     IO,
     Any,
-    BinaryIO,
     Iterable,
     Mapping,
     MutableMapping,
@@ -44,7 +41,7 @@ async def parse_result(response, response_type=None, *, encoding="utf-8"):
             response_type = "text"
             encoding = extras.get("charset", encoding)
         else:
-            raise TypeError("Unrecognized response type: {ct}".format(ct=ct))
+            raise TypeError(f"Unrecognized response type: {ct}")
     if "tar" == response_type:
         what = await response.read()
         return tarfile.open(mode="r", fileobj=BytesIO(what))
@@ -67,7 +64,7 @@ def parse_content_type(ct: str) -> Tuple[str, str, Mapping[str, str]]:
     try:
         main_type, sub_type = pieces[0].split("/")
     except ValueError:
-        msg = 'Invalid mime-type component: "{0}"'.format(pieces[0])
+        msg = f'Invalid mime-type component: "{pieces[0]}"'
         raise ValueError(msg)
     if len(pieces) > 1:
         options = {}
@@ -78,7 +75,7 @@ def parse_content_type(ct: str) -> Tuple[str, str, Mapping[str, str]]:
             try:
                 k, v = opt.split("=", 1)
             except ValueError:
-                msg = 'Invalid option component: "{0}"'.format(opt)
+                msg = f'Invalid option component: "{opt}"'
                 raise ValueError(msg)
             else:
                 options[k.lower()] = v.lower()
@@ -103,7 +100,7 @@ def identical(d1, d2):
             return False
 
         pairs = zip(d1, d2)
-        return all((identical(x, y) for (x, y) in pairs))
+        return all(identical(x, y) for (x, y) in pairs)
 
     return d1 == d2
 
@@ -118,7 +115,7 @@ def human_bool(s) -> bool:
             return True
         if s.lower() in _false_strs:
             return False
-        raise ValueError("Cannot interpret {s!r} as boolean.".format(s=s))
+        raise ValueError(f"Cannot interpret {s!r} as boolean.")
     else:
         return bool(s)
 
@@ -148,11 +145,6 @@ class _DecodeHelper:
 
     def __aiter__(self):
         return self
-
-    # to make it compatible with Python 3.5.0 and 3.5.2
-    # https://www.python.org/dev/peps/pep-0492/#api-design-and-implementation-revisions
-    if sys.version_info <= (3, 5, 2):
-        __aiter__ = asyncio.coroutine(__aiter__)
 
     async def __anext__(self):
         if self._flag:
@@ -190,10 +182,10 @@ def format_env(key, value: Union[None, bytes, str]) -> str:
     if isinstance(value, bytes):
         value = value.decode("utf-8")
 
-    return "{key}={value}".format(key=key, value=value)
+    return f"{key}={value}"
 
 
-def clean_networks(networks: Iterable[str] = None) -> Optional[Iterable[str]]:
+def clean_networks(networks: Optional[Iterable[str]] = None) -> Optional[Iterable[str]]:
     """
     Cleans the values inside `networks`
     Returns a new list
@@ -211,7 +203,7 @@ def clean_networks(networks: Iterable[str] = None) -> Optional[Iterable[str]]:
     return result
 
 
-def clean_filters(filters: Mapping = None) -> str:
+def clean_filters(filters: Optional[Mapping] = None) -> str:
     """
     Checks the values inside `filters`
     https://docs.docker.com/engine/api/v1.29/#operation/ServiceList
@@ -227,7 +219,7 @@ def clean_filters(filters: Mapping = None) -> str:
     return json.dumps(filters)
 
 
-def mktar_from_dockerfile(fileobject: BinaryIO) -> IO:
+def mktar_from_dockerfile(fileobj: Union[BytesIO, IO[bytes]]) -> IO[bytes]:
     """
     Create a zipped tar archive from a Dockerfile
     **Remember to close the file object**
@@ -240,21 +232,21 @@ def mktar_from_dockerfile(fileobject: BinaryIO) -> IO:
     f = tempfile.NamedTemporaryFile()
     t = tarfile.open(mode="w:gz", fileobj=f)
 
-    if isinstance(fileobject, BytesIO):
+    if isinstance(fileobj, BytesIO):
         dfinfo = tarfile.TarInfo("Dockerfile")
-        dfinfo.size = len(fileobject.getvalue())
-        fileobject.seek(0)
+        dfinfo.size = len(fileobj.getvalue())
+        fileobj.seek(0)
     else:
-        dfinfo = t.gettarinfo(fileobj=fileobject, arcname="Dockerfile")
+        dfinfo = t.gettarinfo(fileobj=fileobj, arcname="Dockerfile")
 
-    t.addfile(dfinfo, fileobject)
+    t.addfile(dfinfo, fileobj)
     t.close()
     f.seek(0)
     return f
 
 
 def compose_auth_header(
-    auth: Union[MutableMapping, str, bytes], registry_addr: str = None
+    auth: Union[MutableMapping, str, bytes], registry_addr: Optional[str] = None
 ) -> str:
     """
     Validate and compose base64-encoded authentication header