+++ examples/events.py
@@ -1,36 +1,36 @@
 #!/usr/bin/env python3
 
 import asyncio
+
 from aiodocker.docker import Docker
 from aiodocker.exceptions import DockerError
 
 
 async def demo(docker):
     try:
-        await docker.images.inspect('alpine:latest')
+        await docker.images.inspect("alpine:latest")
     except DockerError as e:
         if e.status == 404:
-            await docker.pull('alpine:latest')
+            await docker.pull("alpine:latest")
         else:
-            print('Error retrieving alpine:latest image.')
+            print("Error retrieving alpine:latest image.")
             return
 
     subscriber = docker.events.subscribe()
 
     config = {
         "Cmd": ["tail", "-f", "/var/log/dmesg"],
-        "Image":"alpine:latest",
-         "AttachStdin": False,
-         "AttachStdout": True,
-         "AttachStderr": True,
-         "Tty": False,
-         "OpenStdin": False,
-         "StdinOnce": False,
+        "Image": "alpine:latest",
+        "AttachStdin": False,
+        "AttachStdout": True,
+        "AttachStderr": True,
+        "Tty": False,
+        "OpenStdin": False,
+        "StdinOnce": False,
     }
-    container = await docker.containers.create_or_replace(
-        config=config, name='testing')
+    container = await docker.containers.create_or_replace(config=config, name="testing")
     await container.start()
-    print("=> created and started container {}".format(container._id[:12]))
+    print(f"=> created and started container {container._id[:12]}")
 
     while True:
         event = await subscriber.get()
@@ -38,21 +38,22 @@ async def demo(docker):
             break
 
         for key, value in event.items():
-            print(key,':', value)
+            print(key, ":", value)
 
         # Demonstrate simple event-driven container mgmt.
-        if event['Actor']['ID'] == container._id:
-            if event['Action'] == 'start':
+        if event["Actor"]["ID"] == container._id:
+            if event["Action"] == "start":
                 await container.stop()
-                print("=> killed {}".format(container._id[:12]))
-            elif event['Action'] == 'stop':
+                print(f"=> killed {container._id[:12]}")
+            elif event["Action"] == "stop":
                 await container.delete(force=True)
-                print("=> deleted {}".format(container._id[:12]))
-            elif event['Action'] == 'destroy':
-                print('=> done with this container!')
+                print(f"=> deleted {container._id[:12]}")
+            elif event["Action"] == "destroy":
+                print("=> done with this container!")
                 break
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     loop = asyncio.get_event_loop()
     docker = Docker()
     try: